requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
langchain-groq
langchain-text-splitters
faiss-cpu
sentence-transformers
pypdf
python-dotenv
streamlit
requests